% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aqi}
\alias{aqi}
\title{Demo data of air quality}
\format{
A data frame with 7140 rows and 6 variables:
\describe{
  \item{Time}{Time for data}
  \item{NO}{Nitric oxide (NO)}
  \item{NO2}{Nitrogen Dioxide (NO2)}
  \item{CO}{Carbon monoxide (CO)}
  \item{SO2}{Sulfur dioxide (SO2)}
  \item{O3}{Ozone (O3)}
}
}
\usage{
aqi
}
\description{
5 days air quality data (1 min resolution) includes: NO, NO2, CO, SO2, O3.
The variables are as follows:
}
\keyword{datasets}
