% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avri.R
\name{avri}
\alias{avri}
\title{Calculate average of variation}
\usage{
avri(
  df,
  bkip,
  cycles = "1 day",
  n = NULL,
  colid = 1,
  st = st,
  et = et,
  na.rm = na.rm,
  wind = wind,
  coliws = 2,
  coliwd = 3
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{bkip}{break input of resampling, such as '1 hour'.}

\item{n}{number of items for each cycle. If n is setted, cycle will invalid.}

\item{colid}{column index of datetime in dataframe.}

\item{st}{start time of resampling. The default value is the fisrt value of datetime column.}

\item{et}{end time of resampling. The default value is the last value of datetime column.}

\item{na.rm}{logical value. Remove NA value or not?}

\item{wind}{logical value. if TRUE, please set coliwd, coliws.}

\item{coliws}{numeric value, colindex of wind speed in dataframe.}

\item{coliwd}{numeric value, colindex of wind direction (degree) in dataframe.}

\item{cycle}{cycle for average variation. By default, cycle is "1 day".}
}
\value{
dataframe for average variation.
}
\description{
Calculates average of variation of time series.
}
\details{
If you have wind data (wind speed, and wind direction in dgree), please set 'wind' as 'TRUE', and set values for 'coliwd' and 'coliws'.
}
