% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm8n.R
\name{dm8n}
\alias{dm8n}
\title{Calculate daily maximum-8-hour ozone}
\usage{
dm8n(df, colid = 1, starthour = 0, endhour = 16, na.rm = TRUE, outputmode = 1)
}
\arguments{
\item{df}{dataframe of time series for ozone.}

\item{colid}{column index for date-time. By default, it equals to 1.}

\item{starthour}{numeric, start hour for calculating 8-hour ozone. By default, it equals to 0.}

\item{endhour}{numeric, end hour for calculating 8-hour ozone. By default, it equals to 16 which means averaging ozone between 16~23.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations?}

\item{outputmode}{numeric, the format of the output, possible value: 1 or 2. See 'value' for the results of filling in 1 or 2.}
}
\value{
a dataframe depends on the value of 'outputMode'. Value 1 will output 1 #' table: maximum-8-hour ozone. Value 1 will output 1 list, which contains 3
tables: 8-hour ozone, statistics of number of valid items within each
calculation interval, and maximum-8-hour ozone.
}
\description{
Calculates daily maximum-8-hour ozone from ozone observation data.
}
\details{
This function can calculate multiple columns of ozone observation data in 1 dataframe with 1 datetime columm (such as ozone concentration in different sites).
}
\examples{
dm8n(aqi[,c(1,6)], colid = 1, starthour = 0, endhour = 16, na.rm = TRUE, outputmode = 2)
}
