% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm.R
\name{fm}
\alias{fm}
\title{format the theme of plot}
\usage{
fm(p, fsz = 13, lsz = 0.5, tkl = 0.2)
}
\arguments{
\item{p}{a ggplot-format plot.}

\item{fsz}{font size in plot.}

\item{lsz}{line size of panel border and axis in plot.}

\item{tkl}{tick length in plot.}
}
\value{
a plot with a new theme.
}
\description{
Format the theme of plot.
}
