% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koh.R
\name{koh}
\alias{koh}
\title{Search kOH value from 'chemspider.com'. Predicted data is generated using the US Environmental Protection Agency’s EPISuite.}
\usage{
koh(spec)
}
\arguments{
\item{spec}{chemical specise to be searched. chemical specise's name or CAS Number is acceptable.}
}
\value{
kOH value
}
\description{
Search kOH value from 'chemspider.com'. Predicted data is generated using the US Environmental Protection Agency’s EPISuite.
}
\examples{
koh("Benzene")
}
