% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{met}
\alias{met}
\title{Demo data of meteorology}
\format{
A data frame with 1287 rows and 5 variables:
\describe{
  \item{Time}{Time for data}
  \item{TEM}{Temperature}
  \item{HUM}{Humidity}
  \item{WS}{Wind speed}
  \item{WD}{Wind direction}
  }
}
\usage{
met
}
\description{
5 days meteorology data (5 mins) includes: Temperature, Humidity,
Wind speed, Wind direction.
The variables are as follows:
}
\keyword{datasets}
