% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsvp.R
\name{nsvp}
\alias{nsvp}
\title{Calculate Surface Area, Volume, Mass of particle by particle number concentration}
\usage{
nsvp(df, dlogdp = FALSE, dsty = 1)
}
\arguments{
\item{df}{dataframe of particle size data: the first column of input is datetime; the other columns are number concentration (N, unit: #/cm3) or log number concentration (dN/dlogdp, unit: #/cm3) for each particle size channel. Column names of the other columns are the middle particle size for each particle size channel.}

\item{dlogdp}{logical value, TRUE if the third column is log number concentration (dN/dlogdp).}

\item{dsty}{numeric value, density of particle namtter.}
}
\value{
a list with 2 dataframe. The first dataframe is a time series for Surface Area (unit: µm2/cm3), Volume (unit: µm3/cm3), Mass  (unit: µg/m3) of each channels; the second dataframe is a time series for total Surface Area, Volume, Mass of all channels.
}
\description{
Calculate Surface Area, Volume, Mass of particle by particle number concentration.
}
