% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofp.R
\name{ofp}
\alias{ofp}
\title{Calculate ozone formation potential}
\usage{
ofp(df, colid = 1)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{colid}{column index of datetime in dataframe.}
}
\value{
dataframes for MIR and OFP.
}
\description{
Calculate ozone formation potential of VOC time series.
}
\details{
Calculate ozone formation potential of VOC time series. Returns MIR dataframe, OFP dataframe.
}
