% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Convert time series into proportion time series}
\usage{
prop(df, cmcase = FALSE)
}
\arguments{
\item{df}{dataframe of time series.}

\item{cmcase}{logical value. Set to TRUE if you only 
want to retain cases which are complete, i.e., have 
no missing values. The default vaule is FALSE.}
}
\value{
a dataframe with proportion time series.
}
\description{
Convert time series into proportion time series.
}
\examples{
prop(voc)
}
