% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setup_tuv}
\alias{setup_tuv}
\title{Demo data of setup for tuv}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{date}{date for each day}
  \item{nt}{data point for each day}
  \item{lat}{lat for each day}
  \item{lon}{lon for each day}
  \item{o3col}{o3 column concentration for each day}
}
}
\usage{
setup_tuv
}
\description{
5 days setup data for tuv includes: nt,	lat,	lon,	o3col.
The variables are as follows:
}
\keyword{datasets}
