% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdf.R
\name{statdf}
\alias{statdf}
\title{Summary of dataframe}
\usage{
statdf(x, n = 2)
}
\arguments{
\item{x}{dataframe of time series.}

\item{n}{digits for reuslt in dataframe.}
}
\value{
a dataframe, columns stands for parameters, rows stands for variables.
}
\description{
Summary of dataframe.
}
\details{
Summary of dataframe: mean, standard deviation (sd), minimum (min), percentiles
(0.25, 0.50, 0.75), maximum (max).
}
\examples{
statdf(aqi)
}
