% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_rs.R
\name{t_rs}
\alias{t_rs}
\title{Resampling time series and return complete time series with new time period.}
\usage{
t_rs(
  df,
  df_date,
  start_time = NULL,
  end_time = NULL,
  break_input,
  na.rm = TRUE
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{df_date}{colname of datetime in.}

\item{start_time}{start time of resampling.}

\item{end_time}{end time of resampling.}

\item{break_input}{interval of resampling.}

\item{na.rm}{logical value. remove NA value or not?}
}
\value{
dataframe with new resolution.
}
\description{
Resampling time series and return complete time series with new time period.
}
