% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trsp.R
\name{trsp}
\alias{trsp}
\title{Resample time series of particle number concentration}
\usage{
trsp(
  df,
  bkip,
  st = NULL,
  et = NULL,
  na.rm = TRUE,
  cpms = TRUE,
  ybk = c(10, 100, 500, 1000),
  nlmt = 20000,
  colsz = 10
)
}
\arguments{
\item{df}{dataframe of time series.  The first column of input is datetime. 
The second column of input is for middle ranges of channels. 
The third column of input is for particle number concentration of each channel at each timepoint.}

\item{bkip}{new resolution breaking input of time series, such as '1 hour'.}

\item{st}{start time of resampling. The default value is the fisrt value of datetime column.}

\item{et}{end time of resampling. The default value is the last value of datetime column.}

\item{na.rm}{logical value. Remove NA value or not?}

\item{cpms}{logical value. Compensate the insufficient amount of the millisecond bit for datetime column.}

\item{ybk}{numeric vector, log breaks of y axis of plot.}

\item{nlmt}{numeric value, uplimit of dNdlogdp colorscales of plot.}

\item{colsz}{numeric value, size of columns in plot.}
}
\value{
a list with 1 dataframe and 1 plot. The dataframe contains a time series with a new time resolution.
}
\description{
Resamples time series, and returns complete time series of of particle number concentration with new time resolution. 
The first column of input is datetime. The second column of input is for middle ranges of channels. 
The third column of input is for particle number concentration of each channel at each timepoint.
}
