% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voc}
\alias{voc}
\title{Demo data of volatile organic compounds (VOCs)}
\format{
A data frame with 120 rows and 9 variables:
\describe{
  \item{Time}{Time for data}
  \item{Propylene}{Propylene}
  \item{Acetylene}{Acetylene}
  \item{n-Butane}{n-Butane}
  \item{trans-2-Butene}{trans-2-Butene}
  \item{2-Methylpentane}{2-Methylpentane}
  \item{3-Methylpentane}{3-Methylpentane}
  \item{2-Methylhexane}{2-Methylhexane}
  \item{Cyclohexane}{Cyclohexane}
}
}
\usage{
voc
}
\description{
5 days VOCs data (1 hour) includes: Propylene, Acetylene, n-Butane,
trans-2-Butene, 2-Methylpentane, 3-Methylpentane, 2-Methylhexane, Cyclohexane.
The variables are as follows:
}
\keyword{datasets}
