% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avri.R
\name{avri}
\alias{avri}
\title{Calculate average of variation}
\usage{
avri(
  df,
  bkip,
  mode = "recipes",
  value = "day",
  colid = 1,
  st = NULL,
  et = NULL,
  na.rm = TRUE,
  digits = 2,
  wind = FALSE,
  coliws = 2,
  coliwd = 3
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{bkip}{new resolution breaking input of time series, such as '1 hour'.}

\item{mode}{for calculating cycles: "recipes", "ncycle", "custom".
"recipes" means using internal setting for calculation.
"ncycle" means setting number of items for per cycle.
"custom" means using 1 column in dataframe as a list of grouping elements for calculation.}

\item{value}{for mode. Possible values for "recipes" are "day", "week", "month", year".
"day" equals to 24 (hours) values in 1 day.
"week" equals to 7 (days) values in 1 week.
"month" equals to 31 (days) values in 1 month.
"year" equals to 12 (months) values in 1 year.
values for "ncycle" is a number representing number of items in per cycle.
values for "custom" is a number representing column index in dataframe.}

\item{colid}{column index of datetime in dataframe.}

\item{st}{start time of resampling. The default value is the fisrt value of datetime column.}

\item{et}{end time of resampling. The default value is the last value of datetime column.}

\item{na.rm}{logical value. Remove NA value or not?}

\item{digits}{numeric value, digits for result dataframe.}

\item{wind}{logical value. if TRUE, please set coliwd, coliws.}

\item{coliws}{numeric value, colindex of wind speed in dataframe.}

\item{coliwd}{numeric value, colindex of wind direction (degree) in dataframe.}
}
\value{
dataframe for average variation.
}
\description{
Calculates average of variation of time series.
}
\details{
If you have wind data (wind speed, and wind direction in dgree), please set 'wind' as 'TRUE', and set values for 'coliwd' and 'coliws'.
}
