% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm8n.R
\name{dm8n}
\alias{dm8n}
\title{Calculate daily maximum-8-hour ozone}
\usage{
dm8n(df, colid = 1, starthour = 0, endhour = 16, na.rm = TRUE, outputmode = 1)
}
\arguments{
\item{df}{dataframe. dataframe for ozone.}

\item{colid}{numeric. Column index for datatime. By default, it equals to 1.}

\item{starthour}{numeric. Start hour for daily calculation. By default, it equals to 0.}

\item{endhour}{numeric.  End hour for daily calculation. By default, it equals to 16 which means averaging ozone between 16~23.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from the calculations?}

\item{outputmode}{numeric. 1 stands for brief results (only DMAX8). 2 stands for detail results (DMAX8, D8, counts for D8). By default, it equal to 1.}
}
\value{
brief results (only DMAX8), or list for for detail results (DMAX8, D8, counts for D8).
}
\description{
Calculates daily maximum-8-hour ozone from ozone observation data.
}
\details{
This function can calculate multiple columns of ozone observation data in 1 dataframe with 1 datetime columm (such as ozone concentration in different sites).
}
