% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_tsw.r
\name{geom_tsw}
\alias{geom_tsw}
\title{Plot time series}
\usage{
geom_tsw(
  df,
  coliws = 2,
  coliwd = 3,
  lsize = 0.8,
  psize = NA,
  msize = 8,
  tickx = 0.05,
  ticky = -0.1,
  tickwd = 270
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{coliws}{column index of wind speed. The default vaule is 2.}

\item{coliwd}{column index of wind direction. The default vaule is 3.}

\item{lsize}{size of line (wind speed). The default vaule is 0.8.}

\item{psize}{size of point (wind speed). The default vaule is NA.}

\item{msize}{size of mark (wind direction). The default vaule is 8.}

\item{tickx}{adjust value for the x position of wind direction legend. The default vaule is 0.05.}

\item{ticky}{adjust value for the y position of wind direction legend. The default vaule is -0.1.}

\item{tickwd}{direction of wind direction legend. The default vaule is 270.}
}
\description{
Easy way to plot time series.
}
\examples{
geom_tsw(met, coliws=4, coliwd=5)
}
