% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koh.R
\name{koh}
\alias{koh}
\title{get kOH}
\usage{
koh(spec)
}
\arguments{
\item{spec}{chemical specise to be searched. chemical specise's name or CAS Number is acceptable.}
}
\value{
kOH value
}
\description{
Searches kOH value from 'chemspider.com'.
}
\details{
Predicted data is generated using the US Environmental Protection Agency’s EPISuite.\cr
Unit is cm3/molecule-sec.\cr
Condition is 25 deg C.
}
\examples{
koh("Benzene")
}
