% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ofp.R
\name{ofp}
\alias{ofp}
\title{Calculate ozone formation potential}
\usage{
ofp(
  df,
  unit = "ppbv",
  t = 25,
  p = 101.325,
  stcd = FALSE,
  sortd = TRUE,
  colid = 1,
  wamg = FALSE
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{unit}{unit for VOC concentration. A character vector from these options: "ugm" or "ppbv". "ugm" means ug/m3. "ppbv" means part per billion volumn.}

\item{t}{Temperature, in Degrees Celsius, used to convert data in 
micrograms per cubic meter to standard conditions 
(25 Degrees Celsius, 101.325 kPa). By default, t equals to 25 Degrees Celsius.}

\item{p}{Pressure, in kPa, used to convert data in micrograms per cubic meter 
to standard conditions (25 Degrees Celsius, 101.325 kPa). 
By default, p equals to 101.325 kPa.}

\item{stcd}{logical. Does it output results in standard conditions? The default vaule is FALSE.}

\item{sortd}{logical value. It determines whether the VOC species
are sorted or not. By default, sortd has value "TRUE".
If TRUE, VOC species in time series will be arranged according to VOC group,
 relative molecular weight, and MIR value.}

\item{colid}{column index for date-time. The default value is 1.}

\item{wamg}{logical. Should warnings be presented? The default vaule is FALSE.}
}
\value{
a list contains 13 tables:
Con_ugm: time series of VOC mass concentration by species;
Con_ugm_mean: the average mass concentration and proportion of VOC by species (sorted from large to small);
Con_ugm_group: time series of VOC mass concentration classified by groups;
Con_ugm_group_mean: the average value and proportion of VOC mass concentration (sorted from large to small) according to major groups;
Con_ppbv: time series of VOC volume concentration by species;
Con_ppbv_mean: the average volume concentration and proportion of VOC by species (sorted from large to small);
Con_ppbv_group: time series of VOC volume concentration according to major groups;
Con_ppbv_group_mean: VOC volume concentration average and proportion (sorted from large to small) according to major groups;
MIR_Result: matched MIR value result;
OFP_Result: OFP time series of VOC by species;
OFP_Result_mean: the average value and proportion of OFP of VOC by species (sorted from large to small);
OFP_Result_group: OFP time series of VOC classified by groups;
OFP_Result_group_mean: the average value and proportion of OFP of VOC according to major groups (sorted from large to small).
}
\description{
Calculate Ozone Formation Potential (OFP) of VOC time series.
}
\details{
The CAS number was matched for each VOC speices (from column name), and the
Maximum Incremental Reactivity (MIR) value was matched through the CAS number and used for time series calculation. \cr
The MIR value comes from "Carter, W. P. (2009). Updated maximum incremental
reactivity scale and hydrocarbon bin reactivities for regulatory applications.
California Air Resources Board Contract, 2009, 339" (revised January 28, 2010).
}
\examples{
ofp(voc)
}
