% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openaq.R
\name{openaq}
\alias{openaq}
\title{Format the data downloaded from OpenAQ}
\usage{
openaq(df)
}
\arguments{
\item{df}{data downloaded from OpenAQ.}
}
\value{
a list which contain a dataframe of formated OpenAQ data and a dataframe for time series of OpenAQ data.
}
\description{
Format the data downloaded from OpenAQ
}
