% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdf.R
\name{statdf}
\alias{statdf}
\title{Summary of dataframe}
\usage{
statdf(df, n = 2, cmcase = FALSE, prop = FALSE)
}
\arguments{
\item{df}{dataframe of time series.}

\item{n}{digits for reuslt in dataframe.}

\item{cmcase}{logical value. Set to TRUE if you only 
want to summary cases which are complete, i.e., have 
no missing values.}

\item{prop}{logical value. Convert time series into 
proportion time series before summary.}
}
\value{
a dataframe, columns stands for parameters, 
rows stands for variables.
}
\description{
Summary of dataframe.
}
\details{
Summary of dataframe: mean, standard deviation (sd), minimum (min), percentiles
(0.25, 0.50, 0.75), maximum (max).
}
\examples{
statdf(voc)
}
