% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statdf.R
\name{statdf}
\alias{statdf}
\title{Summary of dataframe}
\usage{
statdf(x, n = 2)
}
\arguments{
\item{x}{dataframe contains time series.}

\item{n}{digits for reuslt in dataframe.}
}
\value{
dataframe with new resolution.
}
\description{
Summary of dataframe.
}
\details{
Summary of dataframe: mean, sd, min, pecentag evalue (25%, 50%, 70%), max.
}
