% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transp.R
\name{transp}
\alias{transp}
\title{Convert the format of particle size data}
\usage{
transp(
  df,
  prplot = TRUE,
  ybk = c(10, 100, 500, 1000),
  nlmt = 20000,
  colsz = 10
)
}
\arguments{
\item{df}{dataframe of particle size data: a table or a list.}

\item{prplot}{logical. The default vaule is TRUE. If TRUE, plot the data.}

\item{ybk}{numeric vector, log breaks of y axis of plot.}

\item{nlmt}{numeric value, uplimit of dNdlogdp colorscales of plot.}

\item{colsz}{numeric value, size of columns in plot.}
}
\value{
a dataframe. If the input is a table, the output is a list, and if the input is a list, the output is a table.
}
\description{
Converting the format of particle size data. There are 2 types of particle size data: table and list.
For table format: the first column of input is datetime; the other column is the number concentration of each particle size channel, column name is the middle particle size of the particle size channel.  
For list format: the first column of input is datetime. The second column of input is for middle ranges of channels. 
The third column of input is for particle number concentration of each channel at each timepoint.
}
