% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs.R
\name{trs}
\alias{trs}
\title{Resample time series}
\usage{
trs(
  df,
  bkip,
  colid = 1,
  st = NULL,
  et = NULL,
  na.rm = TRUE,
  wind = FALSE,
  coliwd = 2,
  coliws = 3
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{bkip}{break input of resampling, such as '1 hour'.}

\item{colid}{column index of datetime in dataframe.}

\item{st}{start time of resampling. The default value is the fisrt value of datetime column.}

\item{et}{end time of resampling. The default value is the last value of datetime column.}

\item{na.rm}{logical value. Remove NA value or not?}

\item{wind}{logical value. if TRUE, please set coliwd, coliws.}

\item{coliwd}{numeric value, colindex of wind direction (degree) in dataframe.}

\item{coliws}{numeric value, colindex of wind speed in dataframe.}
}
\value{
dataframe with new resolution.
}
\description{
Resamples time series, and returns complete time series with new time resolution.
}
\details{
If you have wind data (wind speed, and wind direction in dgree), please set 'wind' as 'TRUE', and set values for 'coliwd' and 'coliws'.
}
