% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs.R
\name{trs}
\alias{trs}
\title{Resampling time series and return complete time series with new time period.}
\usage{
trs(
  df,
  bkip,
  coli = 1,
  st = NULL,
  et = NULL,
  na.rm = TRUE,
  wind = FALSE,
  coliwd = 2,
  coliws = 3
)
}
\arguments{
\item{df}{dataframe contains time series.}

\item{bkip}{break input of resampling.}

\item{coli}{colindex of datetime in dataframe.}

\item{st}{start time of resampling.}

\item{et}{end time of resampling.}

\item{na.rm}{logical value. Remove NA value or not?}

\item{wind}{logical value. if TRUE, please set coliwd, coliws.}

\item{coliwd}{numeric value, colindex of wind direction (degree) in dataframe.}

\item{coliws}{numeric value, colindex of wind speed in dataframe.}
}
\value{
dataframe with new resolution.
}
\description{
Resampling time series and return complete time series with new time period.
}
