% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuv.R
\name{tuv}
\alias{tuv}
\title{Calculate TUV in batch.}
\usage{
tuv(pathtuv, df, colt)
}
\arguments{
\item{pathtuv}{path for TUV folder.}

\item{df}{dataframe for variable.}

\item{colt}{column index of date.}
}
\value{
result_j dataframe for TUV result.
}
\description{
Calculate TUV in batch.
}
