% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voc}
\alias{voc}
\title{Demo data of volatile organic compounds (VOCs)}
\format{
A data frame with 120 rows and 6 variables:
\describe{
  \item{Time}{Time for data}
  \item{Propylene}{Propylene}
  \item{Acetylene}{Acetylene}
  \item{n.Butane}{n-Butane}
  \item{trans.2.Butene}{trans-2-Butene}
  \item{Cyclohexane}{Cyclohexane}
}
}
\usage{
voc
}
\description{
5 days VOCs data (1 hour) includes: Propylene, Acetylene, n-Butane,
trans-2-Butene, Cyclohexane.
The variables are as follows:
}
\keyword{datasets}
